fix some warnings reduce default window size

+++ plugin/include/QSynthi2/Data.h
@@ -15,7 +15,7 @@ public:
     std::atomic<bool> resetSimulation{false};
 
     // context: simulation thread <-> GUI thread
-    std::atomic<Decimal> barrierType{0};
+    std::atomic<int> barrierType{0};
     std::atomic<Decimal> barrierOffset{0};
     std::atomic<Decimal> barrierWidth{0};
     List<V2> barrierSlits{};
